requirements.txt
openai-whisper
pyannote.audio==3.1.1
python-dotenv
numpy<2.0
tqdm